% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshblock.R
\name{assign_meshblocks_to_sa2}
\alias{assign_meshblocks_to_sa2}
\title{Assign Meshblocks to SA2 Areas}
\usage{
assign_meshblocks_to_sa2(
  meshblock_sf,
  sa2_sf,
  sa2_col = "SA22023_V1",
  output_col = "SA2_CODE"
)
}
\arguments{
\item{meshblock_sf}{Meshblock sf object.}

\item{sa2_sf}{SA2 sf object.}

\item{sa2_col}{Name of SA2 code column in \code{sa2_sf}.}

\item{output_col}{Name of the new meshblock SA2 column.}
}
\value{
Meshblock sf object with SA2 code attached.
}
\description{
Assigns each meshblock to an SA2 using the centroid of the meshblock geometry.
}
