% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{choose_n_samples}
\alias{choose_n_samples}
\title{Choose Adaptive Number of Route Samples}
\usage{
choose_n_samples(
  commuter_count,
  min_samples = 1,
  max_samples = 15,
  commuters_per_sample = 50
)
}
\arguments{
\item{commuter_count}{Number of commuters for one OD pair.}

\item{min_samples}{Minimum number of sampled routes.}

\item{max_samples}{Maximum number of sampled routes.}

\item{commuters_per_sample}{Number of commuters represented by one sampled route.}
}
\value{
Integer number of sampled routes.
}
\description{
Computes the number of representative routes to generate for one OD pair.
}
