% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{fill_na_ids}
\alias{fill_na_ids}
\title{Fill Missing Polygon IDs}
\usage{
fill_na_ids(x)
}
\arguments{
\item{x}{Integer vector.}
}
\value{
Integer vector with missing values filled where possible.
}
\description{
Fills missing IDs in a polygon sequence using neighbouring non-missing IDs.
}
