% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{find_sa2_shapefile}
\alias{find_sa2_shapefile}
\title{Find SA2 Shapefile}
\usage{
find_sa2_shapefile(year = 2023, data_dir = "data")
}
\arguments{
\item{year}{Shapefile year.}

\item{data_dir}{Data directory.}
}
\value{
Path to the SA2 shapefile.
}
\description{
Convenience wrapper for finding Statistical Area 2 shapefiles.
}
