% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_segments.R
\name{make_directional_segments}
\alias{make_directional_segments}
\title{Make Directional Road Segments from One Route}
\usage{
make_directional_segments(route_mat, route_id, weight, direction, digits = 5)
}
\arguments{
\item{route_mat}{Route matrix returned by ghroute.}

\item{route_id}{Route identifier.}

\item{weight}{Route weight.}

\item{direction}{Direction label, either \code{"H2W"} or \code{"W2H"}.}

\item{digits}{Number of decimal digits used when rounding coordinates.}
}
\value{
Data frame of directional road segments.
}
\description{
Splits a route into consecutive directed coordinate segments.
}
