% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_segments.R
\name{traffic_segments_to_sf}
\alias{traffic_segments_to_sf}
\title{Convert Traffic Segments to sf}
\usage{
traffic_segments_to_sf(traffic_df, crs = 4326)
}
\arguments{
\item{traffic_df}{Aggregated segment traffic data frame.}

\item{crs}{Coordinate reference system. Default is WGS84.}
}
\value{
sf object of traffic segments.
}
\description{
Converts an aggregated traffic segment data frame into an sf line object.
}
